requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
matplotlib>=3.4.3
tqdm>=4.62.3
scikit-learn>=0.24.2
Pillow>=8.3.2
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
seaborn>=0.11.2 